@@ -25,21 +26,22 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "@typescript-eslint/parser": "^3.5.0",
-    "css-loader": "^3.6.0",
+    "@typescript-eslint/eslint-plugin": "^3.7.1",
+    "@typescript-eslint/parser": "^3.7.1",
+    "css-loader": "^4.2.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
-    "eslint": "^6.8.0",
+    "eslint": "^7.6.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-react": "^7.20.3",
-    "expect": "^25.5.0",
+    "eslint-plugin-react": "^7.20.5",
+    "expect": "^26.2.0",
     "file-loader": "^6.0.0",
     "jsdom": "^16.2.2",
     "json-loader": "^0.5.4",
     "mini-css-extract-plugin": "^0.9.0",
-    "mocha": "^7.2.0",
+    "mocha": "^8.1.0",
     "nodemon": "^2.0.4",
     "react-test-renderer": "^16.13.1",
     "react-transform-catch-errors": "^1.0.2",
@@ -145,14 +150,14 @@
     "redux-logger": "^3.0.1",
     "redux-saga": "^1.1.3",
     "redux-thunk": "^2.3.0",
-    "regenerator-runtime": "^0.13.5",
+    "regenerator-runtime": "^0.13.7",
     "rimraf": "^3.0.2",
-    "sequelize": "^5.22.3",
-    "sequelize-cli": "^5.5.1",
-    "serialize-javascript": "^3.1.0",
+    "sequelize": "^6.3.4",
+    "sequelize-cli": "^6.2.0",
+    "serialize-javascript": "^4.0.0",
     "spark-md5": "^3.0.1",
     "typescript": "^3.9.6",
-    "webpack": "^4.43.0",
+    "webpack": "^4.44.1",
     "webpack-cli": "^3.3.12"
   }
 }
