@@ -1,8 +1,8 @@
-const fs = require('fs');
+import * as fs from 'fs';
 
 const externalModules = fs.readdirSync('node_modules')
   .filter(x => ['.bin'].indexOf(x) === -1)
   .reduce((acc, cur) => Object.assign(acc, { [cur]: 'commonjs ' + cur }), {});
 
-module.exports = externalModules;
+export default externalModules;
 
