@@ -1,4 +1,4 @@
-const path = require('path');
+import * as path from 'path';
 
 /*
  * __dirname is changed after webpack-ed to another directory
@@ -7,7 +7,7 @@ const path = require('path');
  */
 const CURRENT_WORKING_DIR = process.cwd();
 
-module.exports = {
+export default {
   app: path.resolve(CURRENT_WORKING_DIR, 'app'),
   assets: path.resolve(CURRENT_WORKING_DIR, 'public', 'assets'),
   compiled: path.resolve(CURRENT_WORKING_DIR, 'compiled'),
