@@ -6,23 +6,23 @@
  * NOTE: browser/server is client/server-side rendering respectively in universal/isomorphic javascript
  *
  */
-const PATHS = require('./paths');
-const rules = require('./rules');
-const plugins = require('./plugins');
-const externals = require('./externals');
-const resolve = require('./resolve');
-const UglifyJsPlugin    = require('uglifyjs-webpack-plugin')
+import UglifyJsPlugin from 'uglifyjs-webpack-plugin';
+import externals from './externals';
+import PATHS from './paths';
+import plugins from './plugins';
+import resolve from './resolve';
+import rules from './rules';
 
-module.exports = (env = {}) => {
+export default (env: { browser: boolean }) => {
   const isProduction = process.env.NODE_ENV === 'production';
   const isBrowser = env.browser;
-  console.log(`Running webpack in ${process.env.NODE_ENV} mode on ${isBrowser ? 'browser': 'server'}`);
+  console.log(`Running webpack in ${process.env.NODE_ENV} mode on ${isBrowser ? 'browser' : 'server'}`);
 
   const hotMiddlewareScript = 'webpack-hot-middleware/client?path=/__webpack_hmr&timeout=20000&reload=true';
   const node = { __dirname: true, __filename: true };
 
   const prodServerRender = {
-    mode: "production",
+    mode: 'production',
     devtool: 'source-map',
     context: PATHS.app,
     entry: { server: '../server/index' },
@@ -33,15 +33,15 @@ module.exports = (env = {}) => {
       path: PATHS.compiled,
       filename: '[name].js',
       publicPath: PATHS.public,
-      libraryTarget: 'commonjs2'
+      libraryTarget: 'commonjs2',
     },
     module: { rules: rules({ production: true, browser: false }) },
     resolve,
-    plugins: plugins({ production: true, browser: false })
+    plugins: plugins({ production: true, browser: false }),
   };
 
   const prodBrowserRender = {
-    mode: "production",
+    mode: 'production',
     devtool: 'cheap-module-source-map',
     context: PATHS.app,
     entry: { app: ['./client'] },
@@ -50,7 +50,7 @@ module.exports = (env = {}) => {
       path: PATHS.assets,
       filename: '[name].[chunkhash].js',
       chunkFilename: '[name].[chunkhash:6].js', // for code splitting. will work without but useful to set
-      publicPath: PATHS.public
+      publicPath: PATHS.public,
     },
     optimization: {
       minimizer: [
@@ -59,37 +59,37 @@ module.exports = (env = {}) => {
           uglifyOptions: {
             output: {
               comments: false,
-              ascii_only: true
+              ascii_only: true,
             },
             compress: {
-              warnings: false 
-            }
-          }
-        })
+              warnings: false,
+            },
+          },
+        }),
       ],
       splitChunks: {
-        chunks: "all",
+        chunks: 'all',
         cacheGroups: {
           vendor: {
             test: /node_modules/, // you may add "vendor.js" here if you want to
-            name: "vendor",
-            chunks: "initial",
-            enforce: true
-          }
-        }
+            name: 'vendor',
+            chunks: 'initial',
+            enforce: true,
+          },
+        },
       },
     },
     performance: { /* this to hide 'WARNING in asset size limit: The following asset(s) exceed the recommended size limit (244 KiB).' */
       maxEntrypointSize: 512000,
-      maxAssetSize: 512000
+      maxAssetSize: 512000,
     },
     module: { rules: rules({ production: true, browser: true }) },
     resolve,
-    plugins: plugins({ production: true, browser: true })
+    plugins: plugins({ production: true, browser: true }),
   };
 
   const devBrowserRender = {
-    mode: "development",
+    mode: 'development',
     devtool: 'eval',
     context: PATHS.app,
     entry: { app: ['./client', hotMiddlewareScript] },
@@ -97,18 +97,18 @@ module.exports = (env = {}) => {
     output: {
       path: PATHS.assets,
       filename: '[name].js',
-      publicPath: PATHS.public
+      publicPath: PATHS.public,
     },
     optimization: {
       noEmitOnErrors: true,
     },
     module: { rules: rules({ production: false, browser: true }) },
     resolve,
-    plugins: plugins({ production: false, browser: true })
+    plugins: plugins({ production: false, browser: true }),
   };
 
   const devServerRender = {
-    mode: "development",
+    mode: 'development',
     devtool: 'sourcemap',
     context: PATHS.app,
     entry: { server: '../server/index' },
@@ -123,7 +123,7 @@ module.exports = (env = {}) => {
     },
     module: { rules: rules({ production: false, browser: false }) },
     resolve,
-    plugins: plugins({ production: false, browser: false })
+    plugins: plugins({ production: false, browser: false }),
   };
 
   const prodConfig = isBrowser ? prodBrowserRender : prodServerRender;
