@@ -1,3 +1,3 @@
-export const ENV = process.env.NODE_ENV as ('production' | 'test' | 'development') || 'development' ;
+export const ENV = process.env.NODE_ENV as ('production' | 'test' | 'development') || 'development';
 
 export const GOOGLE_ANALYTICS_ID = process.env.GOOGLE_ANALYTICS_ID || null;
