@@ -1,46 +1,63 @@
-import { Models, sequelize } from '../models';
+import { Request } from 'express';
+import { sequelize } from '../models';
+import User from '../models/users';
 
-const {User} = Models;
+interface GoogleProfile {
+  id: string;
+  displayName: string;
+  _json: {
+    gender: string;
+    picture: string;
+    emails: Array<{ value: string }>
+  }
+}
+
+type DoneFunction = (err: any, user?: any, info?: { message: string }) => void;
 
 /* eslint-disable no-param-reassign */
-function attachGoogleAccount(user, profile, accessToken, done) {
-  user.google = profile.id;
-  user.name = user.name || profile.displayName;
-  user.gender = user.gender || profile._json.gender;
-  user.picture = user.picture || profile._json.picture;
+function attachGoogleAccount(user: User | null, profile: GoogleProfile, accessToken: string, done: DoneFunction) {
+  if (user) {
+    user.google = profile.id;
+    user.name = user.name || profile.displayName;
+    user.gender = user.gender || profile._json.gender;
+    user.picture = user.picture || profile._json.picture;
 
-  return sequelize.transaction((transaction) => user.save({ transaction }).then(() => user.createToken({
-        kind: 'google',
-        accessToken
-      }, { transaction }))).then(() => done(null, user, { message: 'Google account has been linked.' }));
+    return sequelize.transaction((transaction) => user.save({ transaction }).then(() => user.createToken({
+      kind: 'google',
+      accessToken,
+    }, { transaction }))).then(() => done(null, user, { message: 'Google account has been linked.' }));
+  }
+  return done(null, user, { message: 'No such user is found' });
 }
+
 /* eslint-enable no-param-reassign */
 
-function createUserWithToken(profile, accessToken, done) {
+function createUserWithToken(profile: GoogleProfile, accessToken: string, done: DoneFunction) {
   return sequelize.transaction((transaction) => User.create({
-      email: profile._json.emails[0].value,
-      google: profile.id,
-      name: profile.displayName,
-      gender: profile._json.gender,
-      picture: profile._json.picture
-    }, { transaction }).then((user) => user.createToken({
-        kind: 'google',
-        accessToken
-      }, { transaction }).then(() => done(null, user))));
+    email: profile._json.emails[0].value,
+    google: profile.id,
+    name: profile.displayName,
+    gender: profile._json.gender,
+    picture: profile._json.picture,
+  }, { transaction }).then((user) => user.createToken({
+    kind: 'google',
+    accessToken,
+  }, { transaction }).then(() => done(null, user))));
 }
 
 const existingGoogleAccountMessage = [
   'There is already a Google account that belongs to you.',
-  'Sign in with that account or delete it, then link it with your current account.'
+  'Sign in with that account or delete it, then link it with your current account.',
 ].join(' ');
 
 const existingEmailUserMessage = [
   'There is already an account using this email address.',
-  'Sign in to that account and link it with Google manually from Account Settings.'
+  'Sign in to that account and link it with Google manually from Account Settings.',
 ].join(' ');
 
-export default (req, accessToken, refreshToken, profile, done) => User.findOne({
-    where: { google: profile.id }
+export default (req: Request, accessToken: string, refreshToken: string, profile: GoogleProfile, done: DoneFunction) => {
+  User.findOne({
+    where: { google: profile.id },
   }).then((existingUser) => {
     if (req.user) {
       if (existingUser) {
@@ -52,14 +69,15 @@ export default (req, accessToken, refreshToken, profile, done) => User.findOne({
     if (existingUser) return done(null, existingUser);
 
     return User.findOne({
-      where: { email: profile._json.emails[0].value }
+      where: { email: profile._json.emails[0].value },
     }).then((existingEmailUser) => {
       if (existingEmailUser) {
         return done(null, false, { message: existingEmailUserMessage });
       }
       return createUserWithToken(profile, accessToken, done);
     });
-  }).catch((err) => {
+  }).catch((err: Error) => {
     console.log(err);
     return done(null, false, { message: 'Something went wrong trying to authenticate' });
   });
+};
