@@ -1,3 +1,3 @@
-const createStaticAssets = __PRODUCTION__ ? require('./prod') : require('./dev');
+const createStaticAssets = process.env.NODE_ENV === 'production' ? require('./prod') : require('./dev');
 
 export default createStaticAssets;
